Moved connection, subscription and query headset to `Cortex`
